Integrate django-sass-processor with the project (#28)
* Add django-sass-processor

The current front-end workflow requires developers to manually install
libsass first and they need to recompile sass/scss files when there are
new changes.

The addition of this Django app allows us to skip this annoying process
of recompilation. Having a command to precompile all sass/scss files for
the entire project makes it easier for deployment.

libsass and django-compressor are made dev only since
  1. libsass is not required on the production environment, if sass/scss
     files have been precompiled and deployed using offline compilation
  2. django-compressor is required only for offline compilation, when
     using the command `manage.py compilescss`
according to docs.

Signed-off-by: lamcw <thomas@lamcw.com>

* Update README

There is no more need to install libsass and manually compiling css
anymore. Hence the removal of relevant information in README.

Signed-off-by: lamcw <thomas@lamcw.com>